// src/api/models/GeneralOpeningBook.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Opening } from './Opening';
import {
    OpeningFromJSON,
    OpeningFromJSONTyped,
    OpeningToJSON,
    OpeningToJSONTyped,
} from './Opening';
import type { General } from './General';
import {
    GeneralFromJSON,
    GeneralFromJSONTyped,
    GeneralToJSON,
    GeneralToJSONTyped,
} from './General';

/**
 * 
 * @export
 * @interface GeneralOpeningBook
 */
export interface GeneralOpeningBook {
    /**
     * 
     * @type {General}
     * @memberof GeneralOpeningBook
     */
    general: General;
    /**
     * 
     * @type {number}
     * @memberof GeneralOpeningBook
     */
    totalGames: number;
    /**
     * 
     * @type {Array<Opening>}
     * @memberof GeneralOpeningBook
     */
    openings: Array<Opening>;
    /**
     * 
     * @type {number}
     * @memberof GeneralOpeningBook
     */
    otherGameCount: number;
    /**
     * 
     * @type {number}
     * @memberof GeneralOpeningBook
     */
    otherWinCount: number;
}



/**
 * Check if a given object implements the GeneralOpeningBook interface.
 */
export function instanceOfGeneralOpeningBook(value: object): value is GeneralOpeningBook {
    if (!('general' in value) || value['general'] === undefined) return false;
    if (!('totalGames' in value) || value['totalGames'] === undefined) return false;
    if (!('openings' in value) || value['openings'] === undefined) return false;
    if (!('otherGameCount' in value) || value['otherGameCount'] === undefined) return false;
    if (!('otherWinCount' in value) || value['otherWinCount'] === undefined) return false;
    return true;
}

export function GeneralOpeningBookFromJSON(json: any): GeneralOpeningBook {
    return GeneralOpeningBookFromJSONTyped(json, false);
}

export function GeneralOpeningBookFromJSONTyped(json: any, ignoreDiscriminator: boolean): GeneralOpeningBook {
    if (json == null) {
        return json;
    }
    return {
        
        'general': GeneralFromJSON(json['general']),
        'totalGames': json['totalGames'],
        'openings': ((json['openings'] as Array<any>).map(OpeningFromJSON)),
        'otherGameCount': json['otherGameCount'],
        'otherWinCount': json['otherWinCount'],
    };
}

export function GeneralOpeningBookToJSON(json: any): GeneralOpeningBook {
    return GeneralOpeningBookToJSONTyped(json, false);
}

export function GeneralOpeningBookToJSONTyped(value?: GeneralOpeningBook | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'general': GeneralToJSON(value['general']),
        'totalGames': value['totalGames'],
        'openings': ((value['openings'] as Array<any>).map(OpeningToJSON)),
        'otherGameCount': value['otherGameCount'],
        'otherWinCount': value['otherWinCount'],
    };
}